If rsync fails don't delete data

+++ scripts/rsync_rm_loop.sh
@@ -53,7 +53,11 @@ while [ $n_done -lt $n_grids ]; do
             n_done=$(expr $n_done + 1)
             if [ -d "${SMART_job%.txt}_cands" ]; then
                 echo "Syncing ${SMART_job%.txt}_cands"
-                rsync --copy-links -zru ${SMART_job%.txt}_cands pulsar-desktop:~/SMART_cand_sorting/${OBSID}
+                rsync --copy-links -zru ${SMART_job%.txt}_cands prometheus:/data/nswainston/SMART_cand_sorting/${OBSID}
+                if [ $? != 0 ]; then
+                    echo "rsync error exiting"
+                    exit
+                fi
                 echo "Syncing ${SMART_job%.txt}_cands done"
                 echo "Deleting ${SMART_job%.txt}_cands"
                 rm -rf ${SMART_job%.txt}_cands
@@ -68,7 +72,7 @@ while [ $n_done -lt $n_grids ]; do
         # Even if it isn't done start using rysnc update
         if [ -d "${SMART_job%.txt}_cands" ]; then
             echo "Syncing ${SMART_job%.txt}_cands"
-            rsync --copy-links -zru ${SMART_job%.txt}_cands pulsar-desktop:~/SMART_cand_sorting/${OBSID}
+            rsync --copy-links -zru ${SMART_job%.txt}_cands prometheus:/data/nswainston/SMART_cand_sorting/${OBSID}
             echo "Syncing ${SMART_job%.txt}_cands done"
         fi
         cd ..
